Fix Coursier dependency resolution
This is a large commit, as many things had to be changed in order to 
fix Coursier resolution while maintaining a working build.

Coursier did not work in Jest, which was too eager in trying to detect
what the Coursier was doing (it crashed with a message about an 
infinite loop being detected). The test framework has thus changed to
Mocha, run with TypeScript. The types helped actually detect a mistake
in Webpack's official documentation for writing loaders. Additionally,
TypeScript makes it possible to write modern JS without dealing with 
Babel.

Coursier's use of `global.require` was problematic for the way that 
Webpack loads modules. See the comment in `src/main/js/index.js`.
Webpack therefore had to be reintroduced as part of the build.

Finally, the `per-env` package makes it easy to manage build modes.
Setting the `NODE_ENV` environment variable to `production` will build
and test everything in production mode when using the npm scripts. The
fallback if the variable isn't set is `development` mode.