Refactor and fix dependency downloads
Instead of using XHR to get the server response in a String, use 
`fetch` and stream the response to a file. The XHR2 package is still
necessary for Coursier to resolve dependencies.

Rename `FetchDependencies` to `DependencyFetch`. A fetch can either be
a download or a fetch from the cache. Currently, only downloads are
implemented.

Refactor to remove unnecessary abstraction, such as Coursier's `Cache`,
`Fetch`, `EitherT`, etc. The `DependencyCache` is now merged into the
`DependencyFetch` object.

All of these changes speed up the download process (from ~18 seconds to
~9 seconds on my machine), and the downloaded files are no longer
corrupted. They can now correctly be executed.