Fix race condition in command execution
The 'close' event is when the streams close, 'exit' is when the process
exits. The output file may not have been created yet when the streams
are closed, but it will have been created when the process exits. 
Making the promise resolve when the process exits instead of when it
closes therefore fixes a race condition between file creation and read.