Improve error messages of dependency resolution
Previously, when the server responded with something else than a HTTP
200 OK, we still considered the response to be valid. For instance, if
the server replied with 404 Not Found, we still wrote the body of the
HTTP response into a `.pom` file. Coursier would then attempt to parse
the `.pom` file as XML, which would throw a rather cryptic 
`BoxedException` error.

Note that coursier's AlwaysDownload file cache for Scala.js has the
same bug.

This commit fixes the problem by converting HTTP error codes to
internal errors, which can be reported correctly.