Add dependency file stringification
We have a method to parse, so this commit adds a method to stringify.
Note that in general, `stringify(parse(x)) != x` because `x` can
contain the syntactic sugar of `::` and `:::` which looks at the Scala
version in the context to convert to `:` form.

This change will allow us to export dependencies to Bloop later on